Move reservations to method argument
This is the first step in pushing the (impure) effects to the boundary
of the system (AKA Ports and Adapters).

This change necessitates that the reservations array gets added as an
argument to the TryAccept interface definition, which could be
considered a leaky abstraction. This is an example of temporarily making
a code base worse in order to ultimately improve it. In a couple of more
commits, the IMaîtreD interface will be altogether redundant.